+++ waypoints/src/main/kotlin/de/md5lukas/waypoints/config/tracking/PlayerTrackingConfiguration.kt
@@ -1,6 +1,8 @@
 package de.md5lukas.waypoints.config.tracking
 
+import de.md5lukas.konfig.ConfigPath
 import de.md5lukas.konfig.Configurable
+import java.time.Duration
 
 @Configurable
 class PlayerTrackingConfiguration {
@@ -12,6 +14,14 @@ class PlayerTrackingConfiguration {
     private set
 
   var trackingRequiresTrackable = false
+    private set
+
+  @ConfigPath("request.enabled")
+  var requestEnabled = false
+    private set
+
+  @ConfigPath("request.validFor")
+  var requestValidFor: Duration = Duration.ZERO
     private set
 
   var notification = false