+++ waypoints/src/main/kotlin/de/md5lukas/waypoints/util/CommandAPI.kt
@@ -5,4 +5,4 @@ import dev.jorel.commandapi.executors.CommandArguments
 import net.kyori.adventure.text.minimessage.tag.resolver.TagResolver
 
 val CommandArguments.labelResolver: TagResolver
-  get() = "label" placeholder fullInput.substringBefore(' ').removePrefix("/")
+  get() = "label" placeholder fullInput().substringBefore(' ').removePrefix("/")